Allow testrunner to mount local cloned dirs to container
This patch is to enable the user to keep contrail-test and
contrail-test-ci directories (separate) on the host and still get
combined directory structure on the container.

This will allow developers/users to debug or develop on the host and
get the changes right away on the container to test it.